feat: add dashboard control for academic management
- Created a new dashboard for managing academic structures, including sections for subjects, sections, and academic management tools.
- Updated navigation links to redirect users to the new dashboard based on their role.
- Adjusted validation for user and representative IDs to require 7 digits instead of 8.
- Fixed API call in the teacher management module to use the correct user ID field.

+++ backend/src/modules/academic-structure-II/assignments/assignment.model.mjs
@@ -152,7 +152,12 @@ export class AssignmentModel {
     // Desasignar un profesor
     static async unassignTeacher(assignmentId) {
         if (!assignmentId) return { error: 'ID de asignación requerido' };
-
+        // Se verifica si la asignación existe
+        const [existingAssignment] = await db.query(
+            `SELECT * FROM teacher_assignments WHERE assignment_id = ?`,
+            [assignmentId]
+        );
+        if (existingAssignment.length === 0) return { error: 'La asignación especificada no existe.' };
         const [result] = await db.query(
             `DELETE FROM teacher_assignments WHERE assignment_id = ?`,
             [assignmentId]
@@ -162,7 +167,7 @@ export class AssignmentModel {
 
         return { 
             message: 'Asignación eliminada exitosamente',
-            teacher_user_id: result.teacher_user_id 
+            userId: existingAssignment[0].teacher_user_id
         };
     }
 }
